исправление замечаний после первой домашки
