+++ src/view/const.js
@@ -1,0 +1,11 @@
+const TempInfo = {
+  TITLE: 'Amsterdam — Chamonix — Geneva',
+  DATES: '18&nbsp;—&nbsp;20 Mar',
+  COST: '12340'
+};
+
+const FILTER_TYPES = ['everything', 'future', 'present', 'past'];
+
+const SORTING_TYPES = ['day', 'event', 'time', 'price', 'offers'];
+
+export { TempInfo, FILTER_TYPES, SORTING_TYPES };
